Extract system fonts to file
